pyproject.toml
[tool.poetry]
name = "fabric_test"
version = "0.1.0"
description = ""
authors = ["belkanov <78809180+belkanov@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "2.2.10"
djangorestframework = "^3.13.1"
django-environ = "^0.8.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
